chore: Move animations and globals from shared directory
