+++ app/components/Popover.js
@@ -2,7 +2,7 @@
 import * as React from "react";
 import { Popover as ReakitPopover } from "reakit/Popover";
 import styled from "styled-components";
-import { fadeAndScaleIn } from "shared/styles/animations";
+import { fadeAndScaleIn } from "styles/animations";
 
 type Props = {
   children: React.Node,
