+++ app/components/Toast.js
@@ -3,7 +3,7 @@ import { CheckboxIcon, InfoIcon, WarningIcon } from "outline-icons";
 import { darken } from "polished";
 import * as React from "react";
 import styled, { css } from "styled-components";
-import { fadeAndScaleIn, pulse } from "shared/styles/animations";
+import { fadeAndScaleIn, pulse } from "styles/animations";
 import type { Toast as TToast } from "types";
 
 type Props = {
