+++ app/scenes/Document/components/Editor.js
@@ -7,7 +7,7 @@ import { type TFunction, withTranslation } from "react-i18next";
 import styled from "styled-components";
 import breakpoint from "styled-components-breakpoint";
 import { MAX_TITLE_LENGTH } from "shared/constants";
-import { light } from "shared/styles/theme";
+import { light } from "shared/theme";
 import parseTitle from "shared/utils/parseTitle";
 import Document from "models/Document";
 import ClickablePadding from "components/ClickablePadding";
