+++ app/styles/animations.js
