+++ server/emails/components/EmailLayout.js
@@ -1,7 +1,7 @@
 // @flow
 import { Table, TBody, TR, TD } from "oy-vey";
 import * as React from "react";
-import theme from "../../../shared/styles/theme";
+import theme from "../../../shared/theme";
 
 type Props = {
   children: React.Node,
